// core/x-data-grid/src/models/colDef/gridColDef.ts
export type ValueOptions = string | number | { value: any; label: string };