// core/x-data-grid/src/models/params/gridCellParams.ts
export interface GridCellParams<
  R extends GridValidRowModel = any,
  V = unknown,
  F = V,
  N extends GridTreeNode = GridTreeNode,
> {
  /**
   * The grid row id.
   */
  id: GridRowId;
  /**
   * The column field of the cell that triggered the event.
   */
  field: string;
  /**
   * The cell value.
   * If the column has `valueGetter`, use `params.row` to directly access the fields.
   */
  value?: V | undefined;
  /**
   * The cell value formatted with the column valueFormatter.
   */
  formattedValue?: F | undefined;
  /**
   * The row model of the row that the current cell belongs to.
   */
  row: GridRowModel<R>;
  /**
   * The node of the row that the current cell belongs to.
   */
  rowNode: N;
  /**
   * The column of the row that the current cell belongs to.
   */
  colDef: GridStateColDef;
  /**
   * If true, the cell is editable.
   */
  isEditable?: boolean;
  /**
   * The mode of the cell.
   */
  cellMode: GridCellMode;
  /**
   * If true, the cell is the active element.
   */
  hasFocus: boolean;
  /**
   * the tabIndex value.
   */
  tabIndex: 0 | -1;

  async?: boolean;
  registerAPI?: any;
  crud?: EndpointForSelect;
  maxResult?: number;
  typeGet?: number;
  idCountry?: number;
  idState?: number;
  acronym?: boolean;
  multiTypeGet?: number[];
  maxDate?: number;
  minDate?: number;
  useFixedListToAsync?: boolean;
}
// ...
export interface GridRenderEditCellParams<
  R extends GridValidRowModel = any,
  V = any,
  F = V,
  N extends GridTreeNodeWithRender = GridTreeNodeWithRender,
> extends GridCellParams<R, V, F, N>,
  GridEditCellProps<V> {
  /**
   * GridApi that let you manipulate the grid.
   */
  api: GridApiCommunity;
}